+++ Changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Version (NEXT)
+* added quantum kernel methods
+
 ## Version 0.1.0
 * added encodings
 * added training